GotoPage: honour global screen size for pages
When handling page coordinates, irrespective of the setting for
DesktopConfiguration, always use the combined widths and heights for the
attached monitors.

Because pages have to handle mapped windows (that is, unlike switching
desks, windows are not unmapped), when switching pages, windows are
moved in/out of view, depending on the page (rectangle) which is coming
in to view.

In the case of DesktopConfiguration being 'per-monitor', we want to
still use the global screen; those windows on different monitors which
aren't on the monitor we're actively paging for, will still be ignored.

This should go someway to addressing #20.